requirements_repeat.txt
# Requirements for repeat conda environment
# For running live and simulation experiments

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# PsychoPy for experiment control
psychopy>=2025.2.4

# Serial port communication (Biosemi)
pyserial>=3.5
